RSPY-798: Moved on_demand_auxip_staging to auxip_flow.py
