+++ rs_workflows/on_demand_processing.py
@@ -258,72 +258,3 @@ async def on_demand_prip_staging(
 
         # Wait for last task to end (unwrap exceptions if any)
         staged.result()  # type: ignore[unused-coroutine]
-
-
-@flow(name="On-demand Auxip staging")
-async def on_demand_auxip_staging(
-    env: FlowEnvArgs,
-    start_datetime: datetime.datetime | str,
-    end_datetime: datetime.datetime | str,
-    product_type: str,
-    catalog_collection_identifier: str,
-):
-    """
-    Flow to retrieve Auxip files using a ValCover filter with the given time interval defined by
-    start_datetime and end_datetime, select only the type of files wanted if eopf_type is given, stage
-    the files and add STAC items into the catalog.
-    Informations on ValCover filter:
-    https://pforge-exchange2.astrium.eads.net/confluence/display/COPRS/4.+External+data+selection+policies
-
-    Args:
-        env: Prefect flow environment
-        start_datetime: Start datetime for the time interval used to filter the files
-            (select a date or directly enter a timestamp, e.g. "2025-08-07T11:51:12.509000Z")
-        end_datetime: End datetime for the time interval used to filter the files
-            (select a date or directly enter a timestamp, e.g. "2025-08-10T14:00:00.509000Z")
-        product_type: Auxiliary file type wanted
-        catalog_collection_identifier: Catalog collection identifier where CADIP sessions and AUX data are staged
-    """
-
-    # Init flow environment and opentelemetry span
-    flow_env = FlowEnv(env)
-    with flow_env.start_span(__name__, "on-demand-auxip-staging"):
-
-        # Convert datetime inputs to str
-        if isinstance(start_datetime, datetime.datetime):
-            start_datetime = start_datetime.strftime("%Y-%m-%dT%H:%M:%S.%f")[:-3] + "Z"
-        if isinstance(end_datetime, datetime.datetime):
-            end_datetime = end_datetime.strftime("%Y-%m-%dT%H:%M:%S.%f")[:-3] + "Z"
-
-        # CQL2 filter: we use a filter combining a ValCover filter and a product type filter
-        cql2_filter = {
-            "op": "and",
-            "args": [
-                {"op": "=", "args": [{"property": "product:type"}, product_type]},
-                {
-                    "op": "t_contains",
-                    "args": [
-                        {"interval": [{"property": "start_datetime"}, {"property": "end_datetime"}]},
-                        {"interval": [start_datetime, end_datetime]},
-                    ],
-                },
-            ],
-        }
-
-        # Search Auxip products
-        auxip_items = auxip_flow.search_task.submit(
-            flow_env.serialize(),
-            auxip_cql2={"filter": cql2_filter},
-            error_if_empty=False,
-        )
-
-        # Stage Auxip items.
-        staged = staging_task_auxip.submit(
-            flow_env.serialize(),
-            auxip_items,
-            catalog_collection_identifier,
-        )
-
-        # Wait for last task to end.
-        # NOTE: use .result() and not .wait() to unwrap and propagate exceptions, if any.
-        staged.result()  # type: ignore[unused-coroutine]
